+++ R/map_nigeria.R
@@ -2,7 +2,7 @@
 #
 # GPL-3 License
 #
-# Copyright (C) 2019-2020 Victor Ordu.
+# Copyright (C) 2019-2021 Victor Ordu.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -17,7 +17,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>
 
-globalVariables(".")
+globalVariables(c(".", "STATE"))
 
 #' Map of Nigeria
 #'
@@ -40,59 +40,51 @@ globalVariables(".")
 #' @importFrom maps map
 #' @importFrom maps map.text
 #' 
-#' @param region A character vector of a list of Nigerian States to be displayed.
+#' @param region A character vector of regions to be displayed. This could be 
+#' States or Local Government Areas.
 #' @param data An object containing data, principally the variables required to
-#' plotted in a map.
-#' @param x Numeric object or factor (or coercible to one). See \emph{Details}.
-#' @param y Numeric. See \emph{Details}
+#' plot in a map.
+#' @param x,y Numeric object or factor (or coercible to one). See \emph{Details}.
 #' @param breaks Numeric. A vector of length >= 1. If a single value i.e.
-#' scalar, it denote the expected number of breaks. Internally, the function
-#' will attempt to compute approprate category sizes or fail if out-of bounds. 
+#' scalar, it denotes the expected number of breaks. Internally, the function
+#' will attempt to compute appropriate category sizes or fail if out-of bounds. 
 #' Where length is >= 3L, it is expected to be an arithmetic sequence that 
 #' represents category bounds as for \code{\link[base]{cut}} (applicable 
 #' only to choropleth maps).
 #' @param categories The legend for the choropleth-plotted categories. If not 
 #' defined, internally created labels are used.
-#' @param col Colour to be used for the plot. 
-#' @param fill Logical. Whether to colour the plotted map region(s). When 
-#' drawing a choropleth map \code{fill == TRUE} is implied.
-#' @param title Character vector of length 1.
-#' @param caption Character vector of length 1.
-#' @param leg.x Numeric. Position of the legend.
-#' @param leg.y Numeric. Position of the legend.
-#' @param leg.title Character. The legend Title
+#' @param title,caption An optional string for annotating the map.
+#' @param leg.x,leg.y Numeric. Position of the legend.
+#' @param leg.title String. The legend Title
 #' @param leg.orient The orientation of the legend i.e. whether horizontal or
 #' vertical.
-#' @param show.neighbours Logical; \code{TRUE} to display borders of
-#' neighbouring countries.
+#' @param show.neighbours Logical; \code{TRUE} to display the immediate vicinity
+#' neighbouring regions/countries.
 #' @param show.text Logical. Apply labels to the regions of the map.
-#' @param ... Further arguments for function \code{\link[maps]{map}}
+#' @param ... Further arguments passed to \code{\link[maps]{map}}
 #' 
-#' @details The default value for \code{region} is to print all States. 
-#' \code{NULL} will print an outline map, i.e. without internal boundaries.
+#' @details The default value for \code{region} is to print all State boundaries.
 #' \code{data} enables the extraction of data for plotting from an object
 #' of class \code{data.frame}. Columns containing States are identified. The
-#' argument also provided context for quasiquotation when providing the 
-#' \code{x} and {y} arguments.
-#' For \code{x} and \code{y}, when both arguments are supplied, they are taken
-#' to be point coordinates and are plotted as such. If only \code{x} is supplied,
-#' it is presumed that the intention is to make a choropleth map, and thus, 
-#' numeric vector arguments are converted into factors i.e. number classes. 
-#' Otherwise factors or any object that can be coerced to a factor should be used.
-#' For plain plots, the \code{col} argument works the same as with
-#' \code{\link[maps]{map}} and variants. For choropleth maps, the colour
-#' provided represents a (sequential) colour palette based on 
-#' \code{RColorBrewer::brewer.pal}. The possible colour options can be
-#' checked with \code{getOption("choropleth.colours")} and this can also be 
-#' modified by the user.
+#' argument also provides context for quasiquotation when providing the 
+#' \code{x} and \code{y} arguments.
 #' 
-#' @note When adjusting the default colour choiced for choropleth maps, it is
+#' For \code{x} and \code{y}, when both arguments are supplied, they are taken
+#' to be point coordinates. If only \code{x} is supplied, it is presumed that the
+#' intention is to make a choropleth map, and thus, numeric vector arguments are
+#' converted into factors i.e. number classes. Otherwise factors or any object 
+#' that can be coerced to a factor should be used.
+#' 
+#' For plain plots, the \code{col} argument works the same as with
+#' \code{\link[maps]{map}}. For choropleth maps, the colour provided represents 
+#' a (sequential) colour palette based on \code{RColorBrewer::brewer.pal}. The 
+#' available colour options can be checked with 
+#' \code{getOption("choropleth.colours")} and this can also be modified by the user.
+#' 
+#' @note When adjusting the default colour choices for choropleth maps, it is
 #' advisable to use one of the sequential palettes. For a list of of available
 #' palettes, especially for more advanced use, review 
 #' \code{RColorBrewer::display.brewer.all}
-#' 
-#' @return An object of class \code{map}, invisibly; as a side-effect,
-#' results in the drawing of a map of Nigeria.
 #'
 #' @examples
 #' \dontrun{
@@ -112,8 +104,6 @@ map_ng <- function(region = character(),
                    y = NULL,
                    breaks = NULL,
                    categories = NULL,
-                   col = NULL,
-                   fill = FALSE,
                    title = NULL,
                    caption = NULL,
                    show.neighbours = FALSE,
@@ -124,36 +114,45 @@ map_ng <- function(region = character(),
                    leg.orient = c('vertical', 'horizontal'),
                    ...)
 {
-  ## NOTE: In the call to map.text, the name 'database' is actually
-  ## required. This is because internally, there is a call to `eval()`
-  ## which uses its default argument for `envir` i.e. `parent.frame()`.
-  ## An object of any other name is not seen by the quoted call to
-  ## maps::map used by the evaluator function. For more details,
-  ## inspect the source code for `maps::map.text`. This is a bug in the
-  ## `maps` package.
-  
   ## TODO: Allow this function to accept a matrix e.g. for plotting points
-  region <- .processStateParam(region)
-  stopifnot(is.logical(show.neighbours))
+  if (!is.character(region))
+    stop("Expected a character vector as 'region'")
+  region <- .processRegionParam(region)
+  if (!is.logical(show.neighbours))
+    stop("'show.neighbours' should be a boolean")
+  if (length(show.neighbours) > 1L) {
+    warning("Only the first element of 'show.neighbours' was used")
+    show.neighbours <- show.neighbours[1]
+  }
   if (show.neighbours)
-    message("Display of neighbouring countries is disabled")
+    message("Display of neighbouring countries is temporarily disabled")
   value.x <- if (is_null(data) && !is_null(x))
     enquo(x) 
   else 
     enexpr(x)
+  
   chrplth <- if (is_null(value.x) || is_symbol(value.x)) {
-    .validateChoroplethParams(region, data, !!value.x)  # TODO: Refactor
+    .validateChoroplethParams(!!value.x, region, data)  # TODO: Refactor
   } else {
     value.x <- eval_tidy(value.x)
-    .validateChoroplethParams(region, data, value.x)
+    .validateChoroplethParams(value.x, region, data)
   }
   if (!is_null(y))
     chrplth <- FALSE
-  if (is.null(col) && is_false(chrplth))
-    col <- 1L
+  
   database <- .getMapData(region)
-  mapq <- expr(map(database, regions = region, col = col, fill = fill, ...))
+  mapq <- expr(map(database, regions = region, ...))
+  
+  ## Capture 'dots'
+  dots <- list(...)
+  
+  ## Prepare to draw choropleth 
   if (chrplth) {
+    mapq <- expr(map(database, region))  ## TODO: Consider rlang::call2
+    if (!is.null(dots$plot))
+      if (!dots$plot)
+        mapq$plot <- FALSE
+    mapq$fill <- TRUE
     if (!is.null(data)) {
       vl.col <- as_name(value.x)
       st.col <- .regionColumnIndex(data, region)
@@ -171,39 +170,61 @@ map_ng <- function(region = character(),
         breaks = breaks,
         categories = categories
       )
-    cOpts <- .prepareChoroplethOptions(database, cParams, col)
-    col <- cOpts$colors
-    fill <- TRUE
+    cOpts <- .prepareChoroplethOptions(database, cParams, dots$col)
+    mapq$col <- cOpts$colors
     lego <- match.arg(leg.orient)
-    horiz <- if (identical(lego, 'vertical')) FALSE else TRUE
-    leg.tit <- if (!missing(leg.title)) as.character(leg.title)
+    horiz <- if (identical(lego, 'vertical')) 
+      FALSE 
+    else 
+      TRUE
+    leg.tit <- if (!missing(leg.title)) 
+      as.character(leg.title)
   }
-  mp <- if (show.text) {
-    if (!identical(region, 'Nigeria')) {
+  
+  outlineMap <- identical(region, 'Nigeria')
+  
+  ## Draw a `maps::map()` (with or without labels) or capture the object.
+  ##
+  ## NOTE: In the call to map.text, the name 'database' is actually
+  ## required. This is because, internally, there is a call to `eval()`
+  ## which uses its default argument for `envir` i.e. `parent.frame()`.
+  ## An object of any other name is not seen by the quoted call to
+  ## `maps::map()` used by the evaluator function. For more details,
+  ## inspect the source code for `maps::map.text()`. This is actually a 
+  ## bug in the 'maps' package.
+  database <- eval_tidy(mapq)    # possibly, a new `database` is created
+  if (!is.null(dots$plot) && !dots$plot)
+    return(database)
+  if (show.text) {
+    txt <- "Nigeria"
+    if (!outlineMap) {
       txt <- database$name %>%
-        {
-          is <- lapply(region, grep, x = .)
-          ind <- unlist(is)
-          .[ind]
-        } %>%
-        .adjustLabels
+        { 
+          # Account for multi-polygon regions
+          # TODO: Scope this to parent environment
+          rgxRegions <-
+            function(x, nms) {
+              rgx <- paste0("^", x, "(\\:\\d*)?$")
+              grep(rgx, nms, value = TRUE)
+            }
+          is <- lapply(region, rgxRegions, nms = .)
+          unlist(is)
+        }
+      map.text(
+        database,
+        regions = txt,
+        exact = TRUE,                # disallow partial matching
+        labels = .adjustLabels(txt),
+        add = TRUE)
     }
-    map.text(database, region, labels = txt, col = col, fill = fill, ...)
   }
-  else
-    eval_tidy(mapq)
-  if(!is_null(y))
-    if (!.xyWithinBounds(mp, x, y))
+  
+  ## Annotate
+  title(main = title, sub = caption)
+  if(!is_null(y)) {
+    if (!.xyWithinBounds(database, x, y))
       stop("Coordinates are out of bounds of the map")
-  
-  # Capture 'dots' and return visible 
-  # 'map' object if `plot == FALSE`
-  dots <- list(...)
-  params <- names(dots)
-  if ('plot' %in% params)
-    if (is_false(dots$plot))
-      return(mp)
-  
+  }
   if (chrplth) {
     if (is.null(categories))
       categories <- cOpts$bins
@@ -219,8 +240,8 @@ map_ng <- function(region = character(),
   }
   else if(!is_null(y))
     points(x, y, pch = "+") 
-  title(title, caption)
-  invisible(mp)
+  
+  invisible(database)
 }
 
 
@@ -229,22 +250,22 @@ map_ng <- function(region = character(),
 
 
 
-
+## MARK FOR DEPRECATION
+## Processes character input, presumably States, and when empty
+## character vector, provide all the States as a default value.
+## This function's use will be overtaken by the introduction 
+## of LGA level mapping.
 #' @importFrom rlang abort
-.processStateParam <- function(s)
+.processRegionParam <- function(s)
 {
-  if (!identical(s, character()))
-    if (!is.character(s) && !is.null(s))
-      abort("Type of argument supplied to 'region' is invalid.")
+  stopifnot(is.character(s))
+  if (identical(s, "Nigeria"))
+    return(s)
   all.st <- states(all = TRUE)
-  if (is.character(s)) {
-    if (length(s) == 0L)
-      s <- all.st
-    if (!all(s %in% all.st))
-      abort("One or more elements of 'region' is not a Nigerian region")
-  }
-  if (is.null(s))
-    s <- "Nigeria"
+  if (length(s) == 0L)
+    s <- all.st
+  if (!all(s %in% all.st) && !all(s %in% lgas()))
+    abort("One or more elements of 'region' is not a Nigerian region")
   s
 }
 
@@ -256,7 +277,7 @@ map_ng <- function(region = character(),
 #' @importFrom rlang enexpr
 #' @importFrom rlang is_null
 #' @importFrom rlang is_symbol
-.validateChoroplethParams <- function(region = NULL, data = NULL, val = NULL)
+.validateChoroplethParams <- function(val = NULL, region = NULL, data = NULL)
 {
   # TODO: Add some verbosity.
   no.region <- is.null(region) || identical(region, "Nigeria")
@@ -283,19 +304,44 @@ map_ng <- function(region = character(),
 
 
 
+
+
+## S3 Class and methods for internal use:
+.getMapData <- function(x)
+  UseMethod(".getMapData")
+
+
 #' @import mapdata
-#' @importFrom maps SpatialPolygons2map
-.getMapData <- function(region)
+.getMapData.default <- function(x) 
 {
-  stopifnot(is.character(region))
-  if (identical(region, 'Nigeria'))
+  if (identical(x, 'Nigeria'))
     return("mapdata::worldHires")
-  if (!all(is_state(region))) {
-    ss <- paste(region, collapse = ', ')
-    stop("Invalid region(s) for the map: ", ss)
-  }
-  sp <- .getSpatialPolygonsDataFrame()
-  SpatialPolygons2map(sp, namefield = 'admin1Name')
+  .getMapData(states())
+}
+
+
+.getMapData.lgas <- function(x)
+{
+    obj <- shp.lga
+    .getMapFromSpObject(obj)
+ }
+
+
+
+.getMapData.states <- function(x)
+{
+  obj <- shp.state
+  .getMapFromSpObject(obj)
+}
+
+
+
+#' @importFrom maps SpatialPolygons2map
+.getMapFromSpObject <- function(shp.data)
+{
+  stopifnot(inherits(shp.data, 'ShapefileProps'))
+  SpatialPolygons2map(shp.data[['spatialObject']],
+                      namefield = shp.data[['namefield']])
 }
 
 
@@ -303,17 +349,45 @@ map_ng <- function(region = character(),
 
 
 
-
-
-## Read the data from an internal shapefile
-#' @importFrom rgdal readOGR
-.getSpatialPolygonsDataFrame <- function() {
-  dsn <-
-    system.file("extdata/ng_admin", package = 'naijR', mustWork = TRUE)
-  if (identical(dsn, character(1)))
-    stop("The map data could not be found in 'extdata'")
-  readOGR(dsn, .shpLayer, verbose = FALSE)
+.fixBadNames <- function(map)
+{
+  map$names[map$names == "Nasarawa"] <- "Nassarawa"
+  map
 }
+
+
+
+
+# States that are also the names of LGAs
+.LgaLikeStates <-
+  function()
+    c("Bauchi",
+      "Ebonyi",
+      "Ekiti",
+      "Gombe",
+      "Katsina",
+      "Kogi",
+      "Nasarawa",
+      "Oyo")
+
+
+
+## Returns the name currently used by the directory containing the
+## shapefile assets. This is found in inst/extdata.
+.getShapefileDir <- function(region)
+  UseMethod(".getShapefileDir")
+
+.getShapefileDir.states <- function(region)
+{
+    'ng_admin'
+}
+
+.getShapefileDir.lgas <- function(region)
+{
+  'nigeria-lgas'
+}
+
+
 
 
 
@@ -322,13 +396,24 @@ map_ng <- function(region = character(),
 
 
 # For possible export later
-.shpLayer <- "nga_admbnda_adm1_osgof_20161215"
+.shpLayer <- function(level) {
+  if (level == 'state')
+    "nga_admbnda_adm1_osgof_20161215"
+  else if (level == 'lga')
+    "new_lga_nigeria_2003"
+    # "NIGERIA_LGA"
+    # "Nigeria_census_2006_WGS84"
+  else
+    stop("An appropriate layer is not avaiable")
+}
+    
 
 
 
 
 
-
+## Find the index number for the column housing the region names
+## used for drawing a choropleth map
 #' @importFrom rlang abort
 #' @importFrom rlang warn
 .regionColumnIndex <- function(dt, s = NULL)
@@ -336,14 +421,25 @@ map_ng <- function(region = character(),
   stopifnot(is.data.frame(dt))
   if (is.null(s))
     s <- states()
-  n <- vapply(dt, function(x) {
-    if (is.factor(x))
+  
+  ## Checks if a column has the names of States, returning TRUE is so.
+  .fx <- function(x) {
+    if (is.factor(x))    # TODO: Earmark for removal
       x <- as.character(x)
-    if (is.character(x))
-      all(is_state(x))
-    else
-      FALSE
-  }, logical(1))
+    ret <- logical(1)
+    if (is.character(x)) {
+      areStates <- is_state(x)
+      ret <- all(areStates)
+      if (!ret && any(areStates)) {
+        misspelt <- which(!areStates)
+        msg <- sprintf("The following regions are misspelt: %s",
+                       paste(x[!areStates], collapse = ","))
+        warning(msg, call. = FALSE)
+      }
+    }
+    ret
+  }
+  n <- vapply(dt, .fx, logical(1))
   if (!sum(n))
     abort(sprintf("No column with elements in %s.", deparse(substitute(s))))
   if (sum(n) > 1)
@@ -471,32 +567,32 @@ map_ng <- function(region = character(),
 #' @importFrom magrittr %>%
 #' @importFrom rlang abort
 #' @importFrom tools toTitleCase
-.processColouring <- function(color = NULL, n)
+.processColouring <- function(col = NULL, n)
 {
-  .DefaultChoroplethColours <- getOption('choropleth.colours')
-  if (is.null(color))
-    color <- .DefaultChoroplethColours[1]
-  if (is.numeric(color)) {
+  .DefaultChoroplethColours <- getOption('choropleth.colours') # set in zzz.R
+  if (is.null(col))
+    col <- .DefaultChoroplethColours[1]
+  if (is.numeric(col)) {
     default.pal <- .get_R_palette()
     all.cols <- default.pal %>% 
       sub("(green)(3)", "\\1", .) %>% 
       sub("gray", "grey", .)
-    if (!color %in% seq_along(all.cols))
+    if (!col %in% seq_along(all.cols))
       abort(sprintf("'color' must range between 1L and %iL", length(all.cols)))
-    color <- all.cols[color]
+    col <- all.cols[col]
   }
-  among.def.cols <- color %in% .DefaultChoroplethColours
-  in.other.pal <- !among.def.cols && color %in% rownames(brewer.pal.info)
+  among.def.cols <- col %in% .DefaultChoroplethColours
+  in.other.pal <- !among.def.cols && (col %in% rownames(brewer.pal.info))
   pal <-
     if (!among.def.cols) {
       if (!in.other.pal)
         abort(
-          sprintf("'%s' is not a supported colour or palette", color)
+          sprintf("'%s' is not a supported colour or palette", col)
         )
-      color
+      col
     }
     else
-      paste0(tools::toTitleCase(color), "s")
+      paste0(tools::toTitleCase(col), "s")
   RColorBrewer::brewer.pal(n, pal)
 }
 
@@ -572,19 +668,11 @@ map_ng <- function(region = character(),
 
 
 # Rejigs text that is used for labeling a region that has more than 1 polygon
-# NOTE: Quite a bit of hard-coding was used (v. 0.1.0) and should be reviewed
+# TODO: Quite a bit of hard-coding was used (v. 0.1.0) and should be reviewed
 #' @importFrom magrittr %>%
 .adjustLabels <- function(x) 
 {
   stopifnot(is.character(x))
-  .fadj <- function(x, region, pos) {
-    ind <- grep(region, x)
-    finalPos <- ind[pos]
-    x[ind] <- ""
-    x[finalPos] <- region
-    x
-  }
-  
   x <- vapply(
     x,
     FUN.VALUE = character(1L),
@@ -592,13 +680,25 @@ map_ng <- function(region = character(),
       sub("\\:\\d*$", "", l)
   )
   
-  ss <- c("Akwa Ibom", "Cross River")
-  cent <- c(2, 4)
-  for (i in seq_len(2))
-    x <- .fadj(x, ss[i], cent[i])
   
-  dups <- which(duplicated(x))
-  x[dups] <- character(1L)
+  dd <- data.frame(state = c("Akwa Ibom", "Cross River"),
+                   poly = c(2, 4))
+  ss <- dd[["state"]]
+  if (all(grepl(paste0(ss, collapse = "|"), x))) {
+    
+    .fadj <- function(a, reg, pos) {
+      ind <- grep(reg, a)
+      finalPos <- ind[pos]
+      a[ind] <- ""
+      a[finalPos] <- reg
+      a
+    }
+    for (i in seq_len(2))
+      x <- .fadj(x, ss[i], dd[["poly"]][i])
+  }
+  
+  dup <- which(duplicated(x))
+  x[dup] <- character(1L)
   x
 }
 
@@ -606,8 +706,8 @@ map_ng <- function(region = character(),
 
 
 
-# Checks that x and y coordinates are withing the bounds of given map
-# Note: This check is probably too expensive. Conider passing just the range
+# Checks that x and y coordinates are within the bounds of given map
+# Note: This check is probably too expensive. Consider passing just the range
 # though the loss of typing may make this less reliable down the line
 #' @importFrom rlang is_double
 .xyWithinBounds <- function(map, x, y)
@@ -617,4 +717,95 @@ map_ng <- function(region = character(),
   xx <- x >= rr[1] & x <= rr[2]
   yy <- y >= rr[3] & y <= rr[4]
   all(xx, yy)
+}
+
+
+
+
+
+## Get the properties of shape files
+## First collect the path to the shapfile project directory
+## Then, read the shapefile
+## Collect the namefield from the data, based on the region
+#' @importFrom methods slot
+#' @importFrom rgdal readOGR
+ShapefileProps <- function(regions)
+{
+  shp <- .getShapefileDir(regions)
+  dsn <- system.file(file.path("extdata", shp),
+                     package = 'naijR',
+                     mustWork = TRUE)
+  if (identical(dsn, character(1)))
+    stop("The map data could not be found in 'extdata'")
+  pat <- "\\.shp$"
+  shpfl <- list.files(dsn, pat)
+  lyr <- sub(pattern = paste0("(.+)(", pat, ")"), "\\1", shpfl)
+  sp <- readOGR(dsn, lyr, verbose = FALSE)
+  
+  ## The following shapefile, has some unwanted data
+  if (shp == "nigeria-lgas") 
+    sp <- subset(sp, STATE != "Lake")
+
+  dt <- slot(sp, "data")
+  
+  if (shp == "nigeria-lgas")
+    dt$STATE[dt$STATE == "Abuja"] <- "Federal Capital Territory"
+  
+  nmfld <- .fetchNamefield(regions, dt)
+  stopifnot(!is.na(nmfld))
+  new_ShapefileProps(shp, lyr, nmfld, sp)
+}
+
+
+
+
+## Low-level constructor
+new_ShapefileProps <- function(dir, layer, namefield, spObj)
+{
+  structure(list(dir, layer, namefield, spObj),
+            names = c("shapefile", "layer", "namefield", "spatialObject"), 
+            class = "ShapefileProps")
+}
+
+
+
+
+
+
+## Fetch the namefield
+## This is a generic function. The methods we have are for distinguishing
+## how namefields are retrieved, since this varies depending on the kind 
+## of region passed. The interesting case in point is when getting this 
+## field for LGA regions, since the spatial data also contains data on 
+## States. To avoid confusion, when iterating through the data frame, when
+## a column with States is encountered it is skipped as can be seen in the
+## `.fetchNamefield.lgas` method.
+.fetchNamefield <- function(x, ...)
+  UseMethod(".fetchNamefield")
+
+
+.fetchNamefield.lgas <- function(x, dt) {
+  nmfld <- NA
+  for (i in seq_len(ncol(dt))) {
+    if (all(unique(dt[[i]]) %in% states()))  # skip column with States
+      next
+    if (any(x %in% dt[[i]])) {   # just any LGAs will do, since some of them
+      nmfld <- colnames(dt)[i]   # also share names with their State
+      break
+    }
+  }
+  nmfld
+}
+
+
+.fetchNamefield.states <- function(x, dt) {
+  dt[[1]][dt[[1]] == "Nasarawa"] <- "Nassarawa" # Fix for 'ng_admin'
+  nmfld <- NA
+  for (i in seq_len(ncol(dt))) {
+    if (all(x %in% dt[[i]])) {   # all MUST be states
+      nmfld <- colnames(dt)[i]
+      break
+    }
+  }
+  nmfld
 }